+++ common/Storyboarding/Util/OsbSpritePools.cs
@@ -43,10 +43,10 @@ namespace StorybrewCommon.Storyboarding.Util
         public OsbSprite Get(double startTime, double endTime, string path, OsbOrigin origin, bool additive, int poolGroup = 0)
             => Get(startTime, endTime, path, origin, additive ? (sprite, spriteStartTime, spriteEndTime) => sprite.Additive(spriteStartTime, spriteEndTime) : (Action<OsbSprite, double, double>)null, poolGroup);
 
-        public OsbAnimation Get(double startTime, double endTime, string path, int frameCount, int frameDelay, OsbLoopType loopType, OsbOrigin origin = OsbOrigin.Centre, Action<OsbSprite, double, double> finalizeSprite = null, int poolGroup = 0)
+        public OsbAnimation Get(double startTime, double endTime, string path, int frameCount, double frameDelay, OsbLoopType loopType, OsbOrigin origin = OsbOrigin.Centre, Action<OsbSprite, double, double> finalizeSprite = null, int poolGroup = 0)
             => (OsbAnimation)getPool(path, frameCount, frameDelay, loopType, origin, finalizeSprite, poolGroup).Get(startTime, endTime);
 
-        public OsbAnimation Get(double startTime, double endTime, string path, int frameCount, int frameDelay, OsbLoopType loopType, OsbOrigin origin, bool additive, int poolGroup = 0)
+        public OsbAnimation Get(double startTime, double endTime, string path, int frameCount, double frameDelay, OsbLoopType loopType, OsbOrigin origin, bool additive, int poolGroup = 0)
             => Get(startTime, endTime, path, frameCount, frameDelay, loopType, origin, additive ? (sprite, spriteStartTime, spriteEndTime) => sprite.Additive(spriteStartTime, spriteEndTime) : (Action<OsbSprite, double, double>)null, poolGroup);
 
         private OsbSpritePool getPool(string path, OsbOrigin origin, Action<OsbSprite, double, double> finalizeSprite, int poolGroup)
@@ -60,7 +60,7 @@ namespace StorybrewCommon.Storyboarding.Util
             return pool;
         }
 
-        private OsbAnimationPool getPool(string path, int frameCount, int frameDelay, OsbLoopType loopType, OsbOrigin origin, Action<OsbSprite, double, double> finalizeSprite, int poolGroup)
+        private OsbAnimationPool getPool(string path, int frameCount, double frameDelay, OsbLoopType loopType, OsbOrigin origin, Action<OsbSprite, double, double> finalizeSprite, int poolGroup)
         {
             var key = getKey(path, frameCount, frameDelay, loopType, origin, finalizeSprite, poolGroup);
 
@@ -74,7 +74,7 @@ namespace StorybrewCommon.Storyboarding.Util
         private string getKey(string path, OsbOrigin origin, Action<OsbSprite, double, double> action, int poolGroup)
             => $"{path}#{origin}#{action?.Target}.{action?.Method.Name}#{poolGroup}";
 
-        private string getKey(string path, int frameCount, int frameDelay, OsbLoopType loopType, OsbOrigin origin, Action<OsbSprite, double, double> action, int poolGroup)
+        private string getKey(string path, int frameCount, double frameDelay, OsbLoopType loopType, OsbOrigin origin, Action<OsbSprite, double, double> action, int poolGroup)
             => $"{path}#{frameCount}#{frameDelay}#{loopType}#{origin}#{action?.Target}.{action?.Method.Name}#{poolGroup}";
 
         #region IDisposable Support
